+++ hbw/app/javascript/packs/hbw/components/form/static.js.jsx
@@ -3,14 +3,18 @@ import cx from 'classnames';
 import { withConditions, withErrorBoundary } from 'shared/hoc';
 
 modulejs.define('HBWFormStatic', ['React'], (React) => {
-  const HBWFormStatic = ({ params, hidden }) => {
+  const HBWFormStatic = ({
+    params, hidden, task, name, env
+  }) => {
     const className = cx({
       [params.css_class]: true,
       hidden
     }, 'hbw-static');
 
+    const html = env.bpTranslator(`${task.process_key}.${task.key}.${name}`, {}, params.html);
+
     return <div className={className}
-                dangerouslySetInnerHTML={{ __html: params.html }} />;
+                dangerouslySetInnerHTML={{ __html: html }} />;
   };
 
   return compose(withConditions, withErrorBoundary)(HBWFormStatic);