+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v2.7.0 [unreleased]
 - [#575](https://github.com/latera/homs/pull/575) Add custom view for files to order data.
 - [#571](https://github.com/latera/homs/pull/512) Add dynamic `disable_if` and `delete_if` options for select_table.
 - [#580](https://github.com/latera/homs/pull/580) Add several upload file fields for single form.
+- [#584](https://github.com/latera/homs/pull/584) Add locale to field with type `static`.
 
 v2.6.11 [unreleased]
 -------------------